Split TestExecutionManager execution modules
Refactor the oversized parallelization partial into focused source, parallel, and dependency-graph execution modules without changing behavior or APIs.
